Initial commit
PLEASE is a Power Linux Emulated Application Software Environment. It
supports running certain special ppc64 Linux applications within a PASE
environment.

It works by loading and parsing an ELF file. The segments of the ELF
file are loaded at their appropriate address and then the entry point is
jumped to, passing the arguments as appropriate.

Due to limitations in PASE, the text segment cannot have its permissions
changed by userspace to allow replacing it with the ELF image. Instead,
the image must be loaded in the mmap region. Thus, special Linux
programs must be built and only those will run.